Testes unitários dos services de Frequencia e Nota

+++ src/main/java/br/edu/ifal/sigamais/repository/NotaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.ifal.sigamais.model.Nota;
 import java.util.List;
 
-public interface NotaRepository extends JpaRepository<Nota, Long> {
+public interface NotaRepository extends JpaRepository<Nota, Integer> {
     
-    List<Nota> findByMatriculaId(Long matriculaId);
+    List<Nota> findByMatriculaId(Integer matriculaId);
 }
